backend/requirements.txt
fastapi
pydantic
pandas
owslib
requests
uvicorn
mistralai
numpy
faiss-cpu
PyMuPDF